+++ src/mapValues.ts
@@ -1,6 +1,11 @@
 /**
- * Should return new value basing on property name and value. Optionally may return exported REMOVE object to filter out
- * property from the object.
+ * @see {@link mapValues}
+ */
+const UNSET = typeof Symbol !== "undefined" ? Symbol("unset") : {};
+
+/**
+ * A callback function to {@link mapValues}. It should return new value, optionally may return exported mapValuesUnset
+ * value to unset a property in the target the object.
  *
  * @callback MapValuesFn
  * @param {*} value - property value
@@ -8,21 +13,13 @@
  * @example function fn(value, key) { return value * 5; } // all properties will be multiplied
  * @example function fn(value, key) {
  *   if (key === "name") {
- *     return REMOVE;
+ *     return mapValuesUNSET;
  *   }
  *   return value.toUpperCase();
- * }
- * // will upper case all properties and filter out `name` property
- * @returns {*}
+ * } // will upper case all properties and filter out name property
+ * @returns {*} - new value of a property
  */
-const REMOVE = typeof Symbol !== "undefined" ? Symbol("remove") : {};
 type MapValuesFn<S, R> = (value: S[keyof S], key: keyof S) => R;
-
-/**
- * Value that denotes that property should be removed
- *
- * @type {Object}
- */
 
 /**
  * Iterates through object properties returning object with same properties but modified values. Optionally some
@@ -30,11 +27,12 @@ type MapValuesFn<S, R> = (value: S[keyof S], key: keyof S) => R;
  *
  * @param {Object|Array} source - source object
  * @param {MapValuesFn} fn - map function callback that will return new value of a property
- * @example mapValues({ a: 1, b: 2 }, x => x * 2) // will return { a: 1, b: 4 }
- * @example mapValues({ a: 1, b: 2 }, () => REMOVE) // will return {}
+ * @example mapValues({ a: 1, b: 2 }, x => x * 2) // will return { a: 2, b: 4 }
+ * @example mapValues({ a: 1, b: 2 }, () => mapValuesUNSET) // will return {}
+ * @example mapValues({ a: 1, b: 2 }, (value, key) => key === "b" ? mapValuesUNSET : value * 5) // will return { a: 5 }
  * @returns {Object|Array}
  */
-const mapValues = <SourceObject extends object, PossibleReturnValues>(
+const mapValues = <SourceObject extends Record<string, unknown>, PossibleReturnValues>(
     source: SourceObject, fn: MapValuesFn<SourceObject, PossibleReturnValues>,
 ): { [P in keyof SourceObject]: PossibleReturnValues } => {
     const keys = Object.keys(source) as (keyof SourceObject)[];
@@ -45,13 +43,15 @@ const mapValues = <SourceObject extends object, PossibleReturnValues>(
 
     return keys.reduce((result, key) => {
         const value = fn(source[key], key);
-        if (value !== REMOVE) {
+        if (value !== UNSET) {
             result[key] = value; // eslint-disable-line no-param-reassign
         }
         return result;
     }, def);
 };
-mapValues.REMOVE = REMOVE;
 
-export { mapValues };
+export { mapValues, UNSET as mapValuesUNSET };
 
+export type {
+    MapValuesFn,
+};